project/requirements.txt
django==3.2.23
djangorestframework
requests==2.31.0